Split up core API module
