+++ isic/core/api/__init__.py
@@ -1,0 +1,9 @@
+from .collection import CollectionViewSet
+from .image import ImageViewSet
+from .user import user_me
+
+__all__ = [
+    'CollectionViewSet',
+    'ImageViewSet',
+    'user_me',
+]
